% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenes.rn7.R
\name{annotateGenes.rn7}
\alias{annotateGenes.rn7}
\title{annotate genes with reference to rat mRatBN7.2 (rn7).}
\usage{
annotateGenes.rn7(mat, ID.type = "S", full.anno = full.anno.rn7)
}
\arguments{
\item{mat}{data matrix; genes in rows; cell names in columns.}

\item{ID.type}{gene id type: Symbol or Ensemble.}

\item{full.anno}{annotation file for all known genes, automatically loaded in copycat.}
}
\value{
annotations of each genes in rows with chrom and positions.
}
\description{
annotate genes with reference to rat mRatBN7.2 (rn7).
}
\examples{
test.anno.mat <- annotateGenes.rn7(mat=matx, ID.type="ENSEMBLE_id", full.anno = full.anno.rn7)
}
